refactor: (#30) infra.mq 부분 개편
