+++ storties-search/src/main/java/org/example/stortiessearch/application/usecase/CreatePostUseCase.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import org.example.stortiessearch.application.event.CreatePostEvent;
 import org.example.stortiessearch.common.annotations.UseCase;
-import org.example.stortiessearch.infrastructure.mq.producer.CreatePostProducer;
+import org.example.stortiessearch.infrastructure.mq.kafka.event.post.create.CreatePostProducer;
 import org.example.stortiessearch.global.authentication.AuthenticatedUserProvider;
 import org.example.stortiessearch.infrastructure.client.grpc.user.dto.AuthenticatedUser;
 import org.example.stortiessearch.domain.post.CommandPostRepository;
