--- storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/config/KafkaConfig.java
@@ -1,59 +1,0 @@
-package org.example.stortiessearch.infrastructure.mq.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.HashMap;
-import java.util.Map;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.core.ConsumerFactory;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.kafka.listener.ContainerProperties;
-import org.springframework.kafka.support.serializer.ErrorHandlingDeserializer;
-import org.springframework.kafka.support.serializer.JsonDeserializer;
-
-@Configuration
-@EnableKafka
-public class KafkaConfig {
-
-    @Bean
-    public ConsumerFactory<String, Object> consumerFactory(ObjectMapper objectMapper) {
-        JsonDeserializer<Object> jsonDeserializer = new JsonDeserializer<>(objectMapper);
-        jsonDeserializer.addTrustedPackages(
-            "org.example.stortiessearch.infrastructure.mq.consumer",
-            "org.example.stortiessearch.application.event",
-            "*"
-        );
-
-        jsonDeserializer.setRemoveTypeHeaders(false);
-        jsonDeserializer.setUseTypeMapperForKey(false);
-
-        Map<String, Object> props = new HashMap<>();
-        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092"); // 환경변수 추가하기
-        props.put(ConsumerConfig.GROUP_ID_CONFIG, "post-group");
-        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, ErrorHandlingDeserializer.class);
-        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, ErrorHandlingDeserializer.class);
-        props.put(ErrorHandlingDeserializer.KEY_DESERIALIZER_CLASS, StringDeserializer.class);
-        props.put(ErrorHandlingDeserializer.VALUE_DESERIALIZER_CLASS, JsonDeserializer.class);
-
-        return new DefaultKafkaConsumerFactory<>(
-            props,
-            new StringDeserializer(),
-            new ErrorHandlingDeserializer<>(jsonDeserializer)
-        );
-    }
-
-    @Bean(name = "kafkaListenerContainerFactoryWithManualAck")
-    public ConcurrentKafkaListenerContainerFactory<String, Object> kafkaListenerContainerFactoryWithManualAck(
-        ConsumerFactory<String, Object> consumerFactory
-    ) {
-        ConcurrentKafkaListenerContainerFactory<String, Object> factory =
-            new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(consumerFactory);
-        factory.getContainerProperties().setAckMode(ContainerProperties.AckMode.MANUAL_IMMEDIATE);
-        return factory;
-    }
-}
