--- storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/producer/DecreasePostLikeProducer.java
@@ -1,31 +1,0 @@
-package org.example.stortiessearch.infrastructure.mq.producer;
-
-import lombok.RequiredArgsConstructor;
-import org.example.stortiessearch.application.event.DecreasePostLikeEvent;
-import org.example.stortiessearch.infrastructure.mq.KafkaProperties;
-import org.example.stortiessearch.infrastructure.mq.dto.KafkaEvent;
-import org.example.stortiessearch.infrastructure.mq.util.JsonSerializer;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Component;
-
-import static org.example.stortiessearch.infrastructure.mq.KafkaProperties.DECREASE_LIKE_TOPIC;
-
-@Component
-@RequiredArgsConstructor
-public class DecreasePostLikeProducer {
-
-    private final KafkaTemplate<String, KafkaEvent> kafkaTemplate;
-
-    private final JsonSerializer jsonSerializer;
-
-    public void publish(DecreasePostLikeEvent event) {
-        KafkaEvent kafkaEvent = KafkaEvent.builder()
-            .topic(DECREASE_LIKE_TOPIC)
-            .eventClass(DecreasePostLikeEvent.class)
-            .payload(jsonSerializer.toJson(event))
-            .retryCount(0)
-            .build();
-
-        kafkaTemplate.send(DECREASE_LIKE_TOPIC, kafkaEvent);
-    }
-}
