--- storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/util/JsonSerializer.java
@@ -1,39 +1,0 @@
-package org.example.stortiessearch.infrastructure.mq.util;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class JsonSerializer {
-
-    private final ObjectMapper objectMapper;
-
-    public String toJson(Object obj) {
-        try {
-            return objectMapper.writeValueAsString(obj);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("JSON serialization failed", e);
-        }
-    }
-
-    public <T> T fromJson(String json, Class<T> clazz) {
-        try {
-            return objectMapper.readValue(json, clazz);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("JSON deserialization failed", e);
-        }
-    }
-
-    public <T> T fromJson(String json, TypeReference<T> typeRef) {
-        try {
-            return objectMapper.readValue(json, typeRef);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("JSON deserialization failed", e);
-        }
-    }
-}
-
